Fixed ambiguous assertions in junit 4.4

+++ repo/core/branches/core-1.10.x/intact-core/src/test/java/uk/ac/ebi/intact/core/unit/mock/MockDaoFactoryTest.java
@@ -40,7 +40,7 @@ public class MockDaoFactoryTest extends IntactAbstractMockTestCase  {
         ProteinDao proteinDao = IntactContext.getCurrentInstance().getDataContext()
                 .getDaoFactory().getProteinDao();
 
-        assertEquals(75, proteinDao.countUniprotProteins());
+        assertEquals(75, proteinDao.countUniprotProteins().intValue());
     }
 
     @Test(expected = UnsupportedOperationException.class)
